Bringing Fuzzy Finder back, sorta.
So Command-T is definitely where it's at as far as loading new files.
Addresses every gripe I've ever had with Fuzzy Finder and is a lot
faster at indexing file paths, but...

It intends to work with new files, whereas Fuzzy Finder has one
advantage in that it can work with buffers as well.  So, I'm bringing it
back, unbinding it from <Leader>t, and using it for my buffer matching
right along side Command-T for my new file matching.